refactor: move out the component showing selected company to improve rendering
